requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
asgiref==3.7.2
    # via django
build==0.10.0
    # via pip-tools
cfgv==3.3.1
    # via pre-commit
click==8.1.3
    # via pip-tools
distlib==0.3.6
    # via virtualenv
dj-database-url==1.3.0
    # via -r requirements.in
django==4.2
    # via
    #   -r requirements.in
    #   dj-database-url
    #   djangorestframework
djangorestframework==3.14.0
    # via -r requirements.in
filelock==3.12.0
    # via virtualenv
gunicorn==20.1.0
    # via -r requirements.in
identify==2.5.24
    # via pre-commit
nodeenv==1.8.0
    # via pre-commit
packaging==23.1
    # via build
pip-tools==6.13.0
    # via -r requirements.in
platformdirs==3.5.1
    # via virtualenv
pre-commit==3.0.4
    # via -r requirements.in
psycopg2-binary==2.9.6
    # via -r requirements.in
pyproject-hooks==1.0.0
    # via build
python-decouple==3.8
    # via -r requirements.in
pytz==2023.3
    # via djangorestframework
pyyaml==6.0
    # via pre-commit
sqlparse==0.4.4
    # via django
tomli==2.0.1
    # via build
typing-extensions==4.6.3
    # via
    #   asgiref
    #   dj-database-url
virtualenv==20.23.0
    # via pre-commit
wheel==0.40.0
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
